+++ packages/core/system/src/utils/index.ts
@@ -1,2 +1,0 @@
-export { cn } from './cn'
-export { mergeCn } from './classNames'