+++ packages/ui/data-entry/radio/CHANGELOG.md
@@ -9,11 +9,15 @@
 
 ### Patch Changes
 
+- c66e035: Move Collection component to its own package
 - Updated dependencies [0e4b044]
+- Updated dependencies [c66e035]
 - Updated dependencies [f4c9f81]
 - Updated dependencies [bdda654]
+- Updated dependencies [c66e035]
   - @negative-space/field@1.0.0
   - @negative-space/system@1.2.0
+  - @negative-space/collection@2.0.0
 
 ## 1.4.0